requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pyh9f0ad1d_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=24.2.0=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.8.30=hbcca054_0
certifi=2024.8.30=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py312hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.8.5=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
deepdiff=8.0.1=pypi_0
dnspython=2.6.1=pypi_0
docker-pycreds=0.4.0=py_0
docopt-ng=0.9.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
expat=2.6.3=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py312h5eee18b_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
gitdb=4.0.11=pyhd8ed1ab_0
gitignore-parser=0.1.11=pypi_0
gitpython=3.1.43=pyhd8ed1ab_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.11.0=py312h34c39bb_0
hdf5=1.12.1=h2b7332f_3
icu=73.1=h6a678d5_0
idna=3.8=pyhd8ed1ab_0
importlib-metadata=8.5.0=pyha770c72_0
importlib-resources=6.4.5=pypi_0
importlib_metadata=8.5.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.27.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py312h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
jsonpickle=3.3.0=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h06a4308_0
kiwisolver=1.4.4=py312h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.1.7=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.6.82=0
libcurl=8.9.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
lightning-utilities=0.9.0=py312h06a4308_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.9.2=py312h06a4308_0
matplotlib-base=3.9.2=py312h66fe004_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.10=py312h5eee18b_0
mkl_random=1.2.7=py312h526ad5a_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.1.0=pypi_0
munch=4.0.0=pypi_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py312h06a4308_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
omegaconf=2.3.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
orderly-set=5.2.2=pypi_0
packaging=24.1=py312h06a4308_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py312h5eee18b_0
pip=24.2=py312h06a4308_0
platformdirs=4.3.2=pyhd8ed1ab_0
ply=3.11=py312h06a4308_1
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=3.20.3=py312h6a678d5_0
psutil=5.9.0=py312h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pymongo=4.8.0=pypi_0
pyparsing=3.1.4=pypi_0
pyqt=5.15.10=py312h6a678d5_0
pyqt5-sip=12.13.0=py312h5eee18b_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.2=h996f2a0_0
python-dateutil=2.9.0post0=py312h06a4308_2
pytorch=2.4.1=py3.12_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pyyaml=6.0.1=py312h5eee18b_0
pyzmq=25.1.2=py312h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
requests=2.32.3=pyhd8ed1ab_0
rich=13.8.1=pypi_0
sacred=0.8.6=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
seml=0.5.3=pypi_0
sentry-sdk=2.14.0=pyhd8ed1ab_0
setproctitle=1.2.2=py312h5eee18b_0
setuptools=72.1.0=py312h06a4308_0
shellingham=1.5.4=pypi_0
sip=6.7.12=py312h6a678d5_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.2=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torch-geometric=2.5.3=pypi_0
torchaudio=2.4.1=py312_cu121
torchmetrics=1.4.0.post0=py312h06a4308_0
torchtriton=3.0.0=py312
torchvision=0.19.1=py312_cu121
tornado=6.4.1=py312h5eee18b_0
tqdm=4.66.5=py312he106c6f_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.12.5=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
unicodedata2=15.1.0=py312h5eee18b_0
urllib3=2.2.2=pyhd8ed1ab_0
wandb=0.16.6=pyhd8ed1ab_1
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.11.1=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.20.1=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
